*** empty log message ***
git-svn-id: file:///usr/local/svn/fsc2@3525 934aee4a-8510-0410-ae50-ec5b2ea4bb86

+++ trunk/fsc2/src/util.c
@@ -688,7 +688,7 @@ inline long lrnd( double x )
 }
 
 
-inline long irnd( double x )
+inline int irnd( double x )
 {
 	if ( x > LONG_MAX )
 		return LONG_MAX;
